pyproject.toml
[project]
name = "feedback-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.119.0",
    "google-genai>=1.44.0",
    "ipykernel>=7.0.0",
    "nltk>=3.9.2",
    "notebook>=7.4.7",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "transformers>=4.57.1",
    "torch>=2.4.1",
    "uvicorn>=0.37.0",
    "fpdf2>=2.8.5",
    "numpy>=2.2.6",
    "vadersentiment>=3.3.2",
    "matplotlib>=3.10.7",
    "plotly>=6.3.1",
    "umap-learn>=0.5.9.post2",
    "hdbscan>=0.8.40",
    "fastopic>=1.0.1",
    "cerebras>=0.0.1",
]
